Extract itnernals
